XMLHttpRequest/getallresponseheaders-cl.htm

<!doctype html>
<title>Casing of known headers</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
test(() => {
  const client = new XMLHttpRequest
  client.open("GET", "resources/header-content-length.asis", false)
  client.send()
  assert_equals(client.getAllResponseHeaders(), "content-length: 0\r\n")
})
</script>
